move widgte settings into its own file
